+++ NEWS.md
@@ -4,7 +4,7 @@
 
 - Documentation fixes (#384)
 - A bug affecting `route()` function calls when `pbapply` was not installed has been fixed (#386)
-- `oneway()` has been deprecated in favour of faster and easier-to-maintain function `od_oneway()` (also see the in-development `od` package): https://github.com/ropensci/stplanr/pull/387
+- `oneway()` has been deprecated in favour of faster and easier-to-maintain function `od_oneway()` (also see the in-development `od` package) (#387)
 - Various other changes have been made to accomodate the dev version of `dplyr` (#383)
 
 # stplanr 0.5.1
@@ -83,7 +83,7 @@
 ## NEW FEATURES
 
 - New functions `od_aggregate_from()` and `od_aggregate_to()` provide easy ways to aggregate origin-destination pairs. See [#303](https://github.com/ropensci/stplanr/pull/303).
-- Updated `overline2()` is now faster and better documented, thanks to [#307](https://github.com/ropensci/stplanr/pull/307)
+- Updated `overline2()` is now faster and better documented (#307)
 - Updates to `route_dodgr()` function, which provides an interface to the [dodgr](https://github.com/ATFutures/dodgr) package, accepts wider range of inputs
 - Better website and updated function list. See https://ropensci.github.io/stplanr/index.html
 - The `sf` method for `overline()` has been updated so it calls the much faster `overline2()` function